Add default serialization methods and refactor cycler __init__
