+++ axelrod/player.py
@@ -240,49 +240,3 @@ class Player(object):
     @property
     def state_distribution(self):
         return self._history.state_distribution
-
-
-class EvolvablePlayer(Player):
-    """A class for a player that can evolve, for use in the Moran process or with reinforcement learning algorithms.
-
-    This is an abstract base class, not intended to be used directly.
-    """
-
-    name = "EvolvablePlayer"
-
-    def overwrite_init_kwargs(self, **kwargs):
-        """Use to overwrite parameters for proper cloning and testing."""
-        for k, v in kwargs.items():
-            self.init_kwargs[k] = v
-
-    # Serialization and Deserialization to strings.
-
-    def serialize_parameters(self):
-        """Serialize parameters to a string for reinforcement learning."""
-        pass
-
-    @classmethod
-    def deserialize_parameters(cls, serialized):
-        """Deserialize parameters to a Player instance."""
-        pass
-
-    # Optional methods for evolutionary algorithms and Moran processes.
-
-    def mutate(self):
-        """Optional method to allow Player to produce a variant (not in place)."""
-        pass
-
-    def crossover(self, other):
-        """Optional method to allow Player to produce variants in combination with another player. Returns a new
-        Player."""
-        pass
-
-    # Optional methods for particle swarm algorithm.
-
-    def receive_vector(self, vector):
-        """Receive a vector of params and overwrite the Player."""
-        pass
-
-    def create_vector_bounds(self):
-        """Creates the bounds for the decision variables for Particle Swarm Algorithm."""
-        pass
